+++ yandex/data_source_yandex_iot_core_broker.go
@@ -54,6 +54,31 @@ func dataSourceYandexIoTCoreBroker() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"log_options": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"disabled": {
+							Type:     schema.TypeBool,
+							Computed: true,
+						},
+						"log_group_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"folder_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"min_level": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+					},
+				},
+			},
 		},
 	}
 }