+++ yandex/resource_yandex_iot_core_broker.go
@@ -2,6 +2,7 @@ package yandex
 
 import (
 	"fmt"
+	"github.com/yandex-cloud/go-genproto/yandex/cloud/logging/v1"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"google.golang.org/genproto/protobuf/field_mask"
@@ -61,6 +62,37 @@ func resourceYandexIoTCoreBroker() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"log_options": {
+				Type:     schema.TypeList,
+				MaxItems: 1,
+				Optional: true,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"disabled": {
+							Type:     schema.TypeBool,
+							Optional: true,
+						},
+						"log_group_id": {
+							Type:          schema.TypeString,
+							Optional:      true,
+							ConflictsWith: []string{"log_options.0.folder_id"},
+							ExactlyOneOf:  []string{"log_options.0.folder_id", "log_options.0.log_group_id"},
+						},
+						"folder_id": {
+							Type:          schema.TypeString,
+							Optional:      true,
+							ConflictsWith: []string{"log_options.0.log_group_id"},
+							ExactlyOneOf:  []string{"log_options.0.folder_id", "log_options.0.log_group_id"},
+						},
+						"min_level": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -87,12 +119,18 @@ func resourceYandexIoTCoreBrokerCreate(d *schema.ResourceData, meta interface{})
 		return fmt.Errorf("Error getting folder ID while creating IoT Broker: %s", err)
 	}
 
+	logOptions, err := expandBrokerLogOptions(d)
+	if err != nil {
+		return fmt.Errorf("Error expanding log options while creating IoT Registry: %s", err)
+	}
+
 	req := iot.CreateBrokerRequest{
 		FolderId:     folderID,
 		Name:         d.Get("name").(string),
 		Description:  d.Get("description").(string),
 		Labels:       labels,
 		Certificates: certs,
+		LogOptions:   logOptions,
 	}
 
 	op, err := config.sdk.WrapOperation(config.sdk.IoT().Broker().Broker().Create(ctx, &req))
@@ -128,6 +166,9 @@ func flattenYandexIoTCoreBroker(d *schema.ResourceData, broker *iot.Broker) erro
 		return err
 	}
 	d.Set("created_at", getTimestamp(broker.CreatedAt))
+	if logOptions := flattenBrokerLogOptions(broker.LogOptions); logOptions != nil {
+		d.Set("log_options", logOptions)
+	}
 	return nil
 }
 
@@ -194,6 +235,10 @@ func resourceYandexIoTCoreBrokerUpdate(d *schema.ResourceData, meta interface{})
 		updatePaths = append(updatePaths, "labels")
 	}
 
+	if d.HasChange("log_options") {
+		updatePaths = append(updatePaths, "log_options")
+	}
+
 	if len(updatePaths) != 0 {
 		req := iot.UpdateBrokerRequest{
 			BrokerId:    d.Id(),
@@ -201,6 +246,11 @@ func resourceYandexIoTCoreBrokerUpdate(d *schema.ResourceData, meta interface{})
 			Description: d.Get("description").(string),
 			Labels:      labels,
 			UpdateMask:  &field_mask.FieldMask{Paths: updatePaths},
+		}
+
+		req.LogOptions, err = expandBrokerLogOptions(d)
+		if err != nil {
+			return fmt.Errorf("Error expanding log options while updating IoT Registry: %s", err)
 		}
 
 		op, err := config.sdk.IoT().Broker().Broker().Update(ctx, &req)
@@ -245,4 +295,49 @@ func resourceYandexIoTCoreBrokerUpdate(d *schema.ResourceData, meta interface{})
 	d.Partial(false)
 
 	return resourceYandexIoTCoreBrokerRead(d, meta)
+}
+
+func expandBrokerLogOptions(d *schema.ResourceData) (*iot.LogOptions, error) {
+	if v, ok := d.GetOk("log_options.0"); ok {
+		logOptionsMap := v.(map[string]interface{})
+		logOptions := &iot.LogOptions{}
+
+		if disabled, ok := logOptionsMap["disabled"]; ok {
+			logOptions.Disabled = disabled.(bool)
+		}
+		if folderID, ok := logOptionsMap["folder_id"]; ok {
+			logOptions.SetFolderId(folderID.(string))
+		}
+		if logGroupID, ok := logOptionsMap["log_group_id"]; ok {
+			logOptions.SetLogGroupId(logGroupID.(string))
+		}
+		if level, ok := logOptionsMap["min_level"]; ok {
+			if v, ok := logging.LogLevel_Level_value[level.(string)]; ok {
+				logOptions.MinLevel = logging.LogLevel_Level(v)
+			} else {
+				return nil, fmt.Errorf("unknown log level: %s", level)
+			}
+		}
+		return logOptions, nil
+	}
+	return nil, nil
+}
+
+func flattenBrokerLogOptions(logOptions *iot.LogOptions) []interface{} {
+	if logOptions == nil {
+		return nil
+	}
+	res := map[string]interface{}{
+		"disabled":  logOptions.Disabled,
+		"min_level": logging.LogLevel_Level_name[int32(logOptions.MinLevel)],
+	}
+	if logOptions.Destination != nil {
+		switch d := logOptions.Destination.(type) {
+		case *iot.LogOptions_LogGroupId:
+			res["log_group_id"] = d.LogGroupId
+		case *iot.LogOptions_FolderId:
+			res["folder_id"] = d.FolderId
+		}
+	}
+	return []interface{}{res}
 }